Use half total variance in swash
Swash is assumed to be a gaussian distribution with the mean as the wave setup level. In Stockdon, The variable [S] is defined as 4 * standard deviation. This means that S is the elevation difference between extreme low runup and extreme high runup, relative to the setup elevation.

So, as runup is the setup plus the maximum elevation of swash,
The 2% exceedence of runup is calculated by adding S/2 to the setup.

This relationship is shown to have a 99% confidence level in Stockdon et al.